Rework RuntimeFieldType
RuntimeFieldType is currently light and most of its concrete implementation is in AbstractScriptFieldType. Now that they both live in the same module, some members that don't have to do with scripting can be pushed to the base class RuntimeFieldType for other implementors to use.
