+++ server/src/test/java/org/elasticsearch/index/IndexSortSettingsTests.java
@@ -10,12 +10,11 @@ package org.elasticsearch.index;
 
 import org.apache.lucene.search.Query;
 import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.time.DateMathParser;
 import org.elasticsearch.index.fielddata.IndexFieldData;
 import org.elasticsearch.index.fielddata.IndexFieldDataService;
 import org.elasticsearch.index.mapper.MappedFieldType;
 import org.elasticsearch.index.mapper.RuntimeFieldType;
-import org.elasticsearch.index.mapper.ValueFetcher;
 import org.elasticsearch.index.query.SearchExecutionContext;
 import org.elasticsearch.indices.breaker.NoneCircuitBreakerService;
 import org.elasticsearch.indices.fielddata.cache.IndicesFieldDataCache;
@@ -24,6 +23,7 @@ import org.elasticsearch.search.lookup.SearchLookup;
 import org.elasticsearch.search.sort.SortOrder;
 import org.elasticsearch.test.ESTestCase;
 
+import java.time.ZoneId;
 import java.util.Collections;
 import java.util.function.Supplier;
 
@@ -141,9 +141,10 @@ public class IndexSortSettingsTests extends ESTestCase {
         IndicesFieldDataCache cache = new IndicesFieldDataCache(Settings.EMPTY, null);
         NoneCircuitBreakerService circuitBreakerService = new NoneCircuitBreakerService();
         final IndexFieldDataService indexFieldDataService = new IndexFieldDataService(indexSettings, cache, circuitBreakerService, null);
-        MappedFieldType fieldType = new RuntimeFieldType("field", Collections.emptyMap()) {
+        MappedFieldType fieldType = new RuntimeFieldType("field", Collections.emptyMap(), null) {
             @Override
-            public ValueFetcher valueFetcher(SearchExecutionContext context, String format) {
+            protected Query rangeQuery(Object lowerTerm, Object upperTerm, boolean includeLower, boolean includeUpper, ZoneId timeZone,
+                                       DateMathParser parser, SearchExecutionContext context) {
                 throw new UnsupportedOperationException();
             }
 
@@ -160,11 +161,6 @@ public class IndexSortSettingsTests extends ESTestCase {
 
             @Override
             public Query termQuery(Object value, SearchExecutionContext context) {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            protected void doXContentBody(XContentBuilder builder, boolean includeDefaults) {
                 throw new UnsupportedOperationException();
             }
         };
