+++ server/src/test/java/org/elasticsearch/index/mapper/TestRuntimeField.java
@@ -9,10 +9,10 @@
 package org.elasticsearch.index.mapper;
 
 import org.apache.lucene.search.Query;
-import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.time.DateMathParser;
 import org.elasticsearch.index.query.SearchExecutionContext;
 
-import java.io.IOException;
+import java.time.ZoneId;
 import java.util.Collections;
 
 public class TestRuntimeField extends RuntimeFieldType {
@@ -20,17 +20,8 @@ public class TestRuntimeField extends RuntimeFieldType {
     private final String type;
 
     public TestRuntimeField(String name, String type) {
-        super(name, Collections.emptyMap());
+        super(name, Collections.emptyMap(), null);
         this.type = type;
-    }
-
-    @Override
-    protected void doXContentBody(XContentBuilder builder, boolean includeDefaults) throws IOException {
-    }
-
-    @Override
-    public ValueFetcher valueFetcher(SearchExecutionContext context, String format) {
-        return null;
     }
 
     @Override
@@ -41,5 +32,11 @@ public class TestRuntimeField extends RuntimeFieldType {
     @Override
     public Query termQuery(Object value, SearchExecutionContext context) {
         return null;
+    }
+
+    @Override
+    protected Query rangeQuery(Object lowerTerm, Object upperTerm, boolean includeLower, boolean includeUpper,
+                               ZoneId timeZone, DateMathParser parser, SearchExecutionContext context) {
+        throw new UnsupportedOperationException();
     }
 }
