+++ server/src/test/java/org/elasticsearch/indices/IndicesModuleTests.java
@@ -173,7 +173,7 @@ public class IndicesModuleTests extends ESTestCase {
         MapperPlugin plugin = new MapperPlugin() {
             @Override
             public Map<String, RuntimeFieldType.Parser> getRuntimeFieldTypes() {
-                return Map.of("test", (name, node, parserContext) -> null);
+                return Map.of("test", new RuntimeFieldType.Parser((s, parserContext) -> null));
             }
         };
         List<MapperPlugin> plugins = Arrays.asList(plugin, plugin);
@@ -186,7 +186,7 @@ public class IndicesModuleTests extends ESTestCase {
         MapperPlugin plugin = new MapperPlugin() {
             @Override
             public Map<String, RuntimeFieldType.Parser> getRuntimeFieldTypes() {
-                return Map.of(KeywordFieldMapper.CONTENT_TYPE, (name, node, parserContext) -> null);
+                return Map.of(KeywordFieldMapper.CONTENT_TYPE, new RuntimeFieldType.Parser((s, parserContext) -> null));
             }
         };
         List<MapperPlugin> plugins = Collections.singletonList(plugin);
